+++ FrontEnd/src/app/dashboard/articleEdit/[edit]/page.js
@@ -1,0 +1,16 @@
+"use client";
+
+
+const ArticleEdit = ({ params }) => {
+
+
+
+  return (
+    <>
+      <div id="editorjs" className="border-black-500"></div>
+     
+    </>
+  );
+};
+
+export default ArticleEdit;
